separate enum Status and function Filtered Todos
