+++ src/components/TodoFilter/TodoFilter.tsx
@@ -44,7 +44,6 @@ export const TodoFilter: React.FC<Props> = ({
 
       {inputValue && (
         <span className="icon is-right" style={{ pointerEvents: 'all' }}>
-          {/* eslint-disable-next-line jsx-a11y/control-has-associated-label */}
           <button
             data-cy="clearSearchButton"
             type="button"
