+++ src/types/Status.ts
@@ -1,0 +1,5 @@
+export enum Status {
+  all = 'all',
+  active = 'active',
+  completed = 'completed',
+}
